+++ Parser.php
@@ -52,6 +52,9 @@ class Parser
      */
     public function readFeed($file)
     {
+        if (!file_exists($file)) {
+            return false;
+        }
         $feed = file_get_contents($file);
         $json = json_decode($feed, true);
         return $json;
@@ -85,7 +88,7 @@ class Parser
                     $resource = $reader->download($args['source']);
                 }
             } catch (InvalidCertificateException $e) {
-                if ($this->config['silence_security'] === false) {
+                if ($this->config['silence_security'] != 'true') {
                     throw new \Exception($e);
                 }
             }